video/requirements.txt
djangorestframework-simplejwt
djangorestframework
django-rest-auth
markdown
django-filter
social-auth-app-django
django-cors-headers
dj_rest_auth
django-allauth
pillow
gunicorn

boto3
django-storages
psycopg2

django-debug-toolbar

